% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Directory.R
\docType{methods}
\name{Directory}
\alias{$,instance,init}
\alias{$,instance,remove}
\alias{+,Directory,Directory-method}
\alias{+,Directory,character-method}
\alias{+,character,Directory-method}
\alias{-,Directory,character-method}
\alias{/,Directory,character-method}
\alias{Directory}
\alias{as.character,Directory-method}
\alias{show,Directory-method}
\title{Class for working with directories}
\usage{
instance <- Directory(name)
instance$getDirName()
instance$remove(...)

\S4method{show}{Directory}(object)

\S4method{/}{Directory,character}(e1, e2)

\S4method{-}{Directory,character}(e1, e2)

\S4method{+}{Directory,character}(e1, e2)

\S4method{+}{character,Directory}(e1, e2)

\S4method{+}{Directory,Directory}(e1, e2)

\S4method{as.character}{Directory}(x, ...)
}
\arguments{
\item{...}{arguments passed to init}

\item{object}{instance of Directory}

\item{e1}{lhs}

\item{e2}{rhs}

\item{name}{character of directory}
}
\description{
Class for working with directories
}

