% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLocalPackageUpdates.R
\name{getLocalPackageUpdates}
\alias{getLocalPackageUpdates}
\title{Function to report differences in NEWS file between CRAN Verion and local version of a package.}
\usage{
getLocalPackageUpdates(package)
}
\arguments{
\item{package}{as character}
}
\description{
Function to report differences in NEWS file between CRAN Verion and local version of a package.
}

