% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/init_thesis.R
\name{init_thesis}
\alias{init_thesis}
\title{Add R/Latex Project structure for article/thesis}
\usage{
init_thesis(dir = ".")
}
\arguments{
\item{dir}{character}
}
\description{
Adds a couple of files/folders which are helpful for reproducible articles with RMarkdown/Latex/R
}
\details{
This structure expects \code{pandoc}, \code{make} and \code{latex} to be installed.In Windows should add those programs to your \code{PATH} variable plus \code{RScript}.
}

