% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/watch.R
\name{watch}
\alias{getPandocPath}
\alias{make}
\alias{setPathVariable}
\alias{watch}
\title{Continously run make}
\usage{
watch(path = ".", nodosfilewarning = TRUE)

make(..., path = ".")

setPathVariable(...)

getPandocPath()
}
\arguments{
\item{path}{character}

\item{nodosfilewarning}{logical}

\item{...}{characters passed as arguments in \code{\link{system2}}}
}
\description{
Continously run make

Run make in a folder
}

