% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/watch.R
\name{watch}
\alias{watch}
\title{Continously run make}
\usage{
watch(path = ".", nodosfilewarning = TRUE)
}
\arguments{
\item{path}{character}

\item{nodosfilewarning}{logical}
}
\description{
Continously run make
}

