% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/watch.R
\name{watch}
\alias{make}
\alias{watch}
\title{Continously run make}
\usage{
watch(path = ".", nodosfilewarning = TRUE)

make(..., path = ".")
}
\arguments{
\item{path}{character}

\item{nodosfilewarning}{logical}

\item{path}{character}
}
\description{
Continously run make

Run make in a folder
}

